resolve get data
